+++ lib/src/common.dart
@@ -149,20 +149,25 @@ abstract class PluginCommand extends Command<Null> {
   ///    "client library" package, which declares the API for the plugin, as
   ///    well as one or more platform-specific implementations.
   Stream<Directory> _getAllPlugins() async* {
-    final Set<String> packages = new Set<String>.from(argResults[_pluginsArg]);
+    final Set<String> plugins = new Set<String>.from(argResults[_pluginsArg]);
     await for (FileSystemEntity entity
         in packagesDir.list(followLinks: false)) {
       // A top-level Dart package is a plugin package.
       if (_isDartPackage(entity)) {
-        if (packages.isEmpty || packages.contains(p.basename(entity.path))) {
+        if (plugins.isEmpty || plugins.contains(p.basename(entity.path))) {
           yield entity;
         }
       } else if (entity is Directory) {
         // Look for Dart packages under this top-level directory.
         await for (FileSystemEntity subdir in entity.list(followLinks: false)) {
           if (_isDartPackage(subdir)) {
-            if (packages.isEmpty ||
-                packages.contains(p.basename(subdir.path))) {
+            // If --plugin=my_plugin is passed, then match all federated
+            // plugins under 'my_plugin'. Also match if the exact plugin is
+            // passed.
+            if (plugins.isEmpty ||
+                plugins.contains(
+                    p.relative(subdir.path, from: packagesDir.path)) ||
+                plugins.contains(p.basename(entity.path))) {
               yield subdir;
             }
           }